+++ frontend/src/app/main/ui/workspace/sidebar/options/menus/typography.cljs
@@ -360,11 +360,8 @@
   [{:keys [values on-change on-blur]}]
   (let [{:keys [line-height
                 letter-spacing]} values
-
         line-height (or line-height "1.2")
         letter-spacing (or letter-spacing "0")
-        line-height-nillable (if (= (str line-height) "1.2") false true)
-
         handle-change
         (fn [value attr]
           (on-change {attr (str value)}))]
@@ -379,11 +376,11 @@
        {:min -200
         :max 200
         :step 0.1
-        :default "1.2"
+        :default-value "1.2"
         :class (stl/css :line-height-input)
         :value (attr->string line-height)
-        :placeholder (tr "settings.multiple")
-        :nillable line-height-nillable
+        :placeholder (if (= :multiple line-height) (tr "settings.multiple") "--")
+        :nillable (= :multiple line-height)
         :on-change #(handle-change % :line-height)
         :on-blur on-blur}]]
 
@@ -397,10 +394,12 @@
        {:min -200
         :max 200
         :step 0.1
+        :default-value "0"
         :class (stl/css :letter-spacing-input)
         :value (attr->string letter-spacing)
-        :placeholder (tr "settings.multiple")
+        :placeholder (if (= :multiple letter-spacing) (tr "settings.multiple") "--")
         :on-change #(handle-change % :letter-spacing)
+        :nillable (= :multiple letter-spacing)
         :on-blur on-blur}]]]))
 
 (mf/defc text-transform-options